requirements.txt
#Usar Python 3.10

# Core do LangChain
langchain
langchain-ollama
langchain-experimental
langchain-community
langchainhub
langchain-google-genai

# Integrações antigas removidas:
#   - langchain_openai
#   - openai

# Outras bibliotecas
pinecone-client==4.1.2
PyMuPDF==1.24.7
wikipedia==1.4.0
numpy==1.26.4
streamlit==1.36.0
numexpr==2.10.1
pandas-datareader==0.10.0
duckduckgo_search==6.1.12
scikit-learn==1.5.1
matplotlib==3.9.1
ipywidgets==8.1.3
tqdm==4.66.4
ipykernel==6.29.5
PyYAML==6.0.1
pydantic-settings
